requirements.txt
# Voice2Marketing - 음성 기반 소상공인 마케팅 자동 생성 서비스
# 필수 의존성

# Web Framework
fastapi>=0.121.0
uvicorn[standard]>=0.38.0
python-multipart>=0.0.20

# Database & ORM
sqlalchemy>=2.0.0

# Authentication & Security
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0

# AI & ML
openai>=2.7.1

# Utilities
python-dotenv>=1.2.1
requests>=2.31.0

